requirements.txt
groq
python-dotenv
keyring
sentence-transformers
numpy
requests
beautifulsoup4
fastapi
uvicorn
